Changed transaction model to use long instead of int.

+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/qr/models/QRTransaction.java
@@ -3,12 +3,12 @@ package nl.tudelft.cs4160.trustchain_android.qr.models;
 import com.squareup.moshi.Json;
 
 public class QRTransaction {
-    public int down;
-    public int up;
+    public long down;
+    public long up;
 
     @Json(name = "total_up")
-    public int totalUp;
+    public long totalUp;
 
     @Json(name = "total_down")
-    public int totalDown;
+    public long totalDown;
 }
